TNM-4929 Create a client package in the OCPP-J API subproject
