+++ ocpp-j-api/src/main/scala/com/thenewmotion/ocpp/json/api/client/OcppJsonClient.scala
@@ -1,11 +1,10 @@
 package com.thenewmotion.ocpp
-package json
-package api
+package json.api
+package client
 
 import java.net.URI
-
-import scala.concurrent.ExecutionContext
 import javax.net.ssl.SSLContext
+import scala.concurrent.ExecutionContext
 
 /**
  * An OCPP-J client implemented using Java-WebSocket.
